Next step

+++ pw/branches/r1117/Src/PW_Client/SelectGameModeScreen.cpp
@@ -164,12 +164,12 @@ void SelectGameModeScreen::Step( bool bAppActive )
 
   // 3. Select hero
   if (locked->GetLobbyStatus() == lobby::EClientStatus::InCustomLobby && g_sessionStatus == WebLauncherPostRequest::RegisterInSessionRequest_Joined) {
-    int heroId = std::min(std::max((size_t)(g_playerHeroId - 1), 0u), _countof(heroes) - 1u);
+    int heroId = std::min(std::max((size_t)(size_t)(g_playerHeroId - 1), size_t(0)), size_t(_countof(heroes) - 1));
     locked->ChangeCustomGameSettings(lobby::ETeam::Enum(g_playerTeamId), lobby::ETeam::Enum(g_playerTeamId), heroes[heroId]);
     g_sessionStatus = WebLauncherPostRequest::RegisterInSessionRequest_HeroSelected;
   }
   if ((locked->GetLobbyStatus() == lobby::EClientStatus::InCustomLobby || g_localGameRun) && g_sessionStatus == WebLauncherPostRequest::RegisterInSessionRequest_WebJoined) {
-    int heroId = std::min(std::max((size_t)(g_playerHeroId - 1), 0u), _countof(heroes) - 1u);
+    int heroId = std::min(std::max((size_t)(size_t)(g_playerHeroId - 1), size_t(0)), size_t(_countof(heroes) - 1));
     locked->ChangeCustomGameSettings(lobby::ETeam::Enum(g_playerTeamId), lobby::ETeam::Enum(g_playerTeamId), heroes[heroId]);
     locked->SetDeveloperParty(g_playerPartyId);
     g_sessionStatus = WebLauncherPostRequest::RegisterInSessionRequest_WebHeroSelected;
