+++ pw/branches/r1117/Src/Server/WebClient/WebClient.cpp
@@ -118,7 +118,7 @@ void CurlClient::InitCurl()
 
   if ( s_debugCurl ) {
     curl_easy_setopt( _curl, CURLOPT_VERBOSE, (LONG)1 );
-    curl_easy_setopt( _curl, CURLOPT_DEBUGFUNCTION, (LONG)CurlCb_debug_callback );
+    curl_easy_setopt( _curl, CURLOPT_DEBUGFUNCTION, CurlCb_debug_callback );
   }
 
   curl_easy_setopt( _curl, CURLOPT_NOPROGRESS, 1L );
